build: :construction_worker: Improve the CubeMX autogeneration code. (#111)
* build: :construction_worker: Improve the CubeMX autogeneration code.

Provide a build option to bypass CubeMX generation (set AUTOGEN_CUBEMX=OFF in the top-level make command). Change method for appending custom targets to the cubemx Makefile to use a ile copy instead of string printing.

* build: :construction_worker: Enable building indepdent of cubemx generation.

Instead of appending targets to the CubeMX makefile, the targets are appended to a different copy ("CustomMakefile.mk"). This lets us leverage Make's prerequisite solving to independently generate the code and append the targets as necessary.