refactor: Remove AnalogInput class. Update status light. Send diagnostics over CAN.
The AnalogInput class only wrapped an ADC peripheral with a moving average and a mapping function. Moved the behaviour to individual methods in main.cc to simplify it. Didn't keep the moving average - we will handle this in the control system if deemed necessary, and I never liked how we used a 20step moving average which delays the reponse time by 100/2=50 ms. Unhandled Simulink outputs are now sent over CAN or properly consumed.