TMS Bringup (#344)
* Minimal TMS CAN update

* refactor: Simplify TMS Temperature Sensors

Remove the TempSensorManager class since all it did was run a loop over the temp sensors. This loop is now in main. Moves the temperature sensor calculations to a separate header file. Simplifies the TemperatureSensor class.

* refactor: Remove BmsBroadcaster Object

Create the BmsBroadcast message in a standalone function so that we don't need to maintain another class. Moves the avg / low / high temperature calculations into this Pack() method to cut down the number of arguments.

* refactor: Update FanController. Remove OS from TMS.

The step size in FanController is now determined by the task rate. FreeRTOS has been disabled. There was only a single task so an RTOS was unnecessary complexity.

* Port TMS to cli

* Formatting