Scope AmkState within class. Remove constructor.
The AmkManager class used to accept a `starting_state` constructor arg
to accelerate testing. However, since the output is dependent on
stepping through previous states, this constructor can cause the output
to differ.

Removed the constructor, but then had to update the tests. Created a
`CycleToState(starting_state)` method to replace the functionality.