Remove debugging transitions.
We temporarily bypassed the FSM transitions based on CAN messages since
the LV wasn't receiving CAN messages, but this was fixed in
232f6e3566497dbc04666a4dbc3e95214aa55fc6
by enabling the RX0 CAN interrupt.

Removes a warning in the LvControllerStatus message by making `.elapsed`
a signed integer (equally, could have made the datatype a uint32 in
code).