Rename Battery Monitor precharge states.
The `INIT_PRECHARGE` and `PRECHARGE` states had misleading names in the
FSM. The names were 1 state off:
`PRECHARGE` _was_ when all three contactors are closed, but this is the
end of the precharge sequence and the precharge contactor is about to
open to enter the `RUNNING` state.
`INIT_PRECHARGE` _was_ when all NEG & PRE are closed, POS is open. This
is the precharging state.

The labels have been adjusted
`PRECHARGE` -> `PRECHARGE_DONE` and
`INIT_PRECHARGE` -> `PRECHARGE`
to make the code reflect the actual behaviour.