Prevent immediate transition BM RUNNING state.
When testing the sequence on the physical battery, the precharge
sequence would finish but the contactors would immediately open.

We traced the cause to the "RUNNING" status exit condition which did not
include enough time for the BMS to respond to the contactor command, so
the contactors did not immediately match the desired state and the
system "faulted". Adds a 500ms delay to prevent this.