Consolidate Controls Modules for FrontController (#400)
* Move C++ Modules into FrontController project

* Rename inc/ to disambiguate

* Consolidate enums

* Update AMKBlock to use new CAN message classes

* Fix incorrect enums

* Create a temporary Makefile test system

* Remove unused enum values

* Scope AmkState within class. Remove constructor.

The AmkManager class used to accept a `starting_state` constructor arg
to accelerate testing. However, since the output is dependent on
stepping through previous states, this constructor can cause the output
to differ.

Removed the constructor, but then had to update the tests. Created a
`CycleToState(starting_state)` method to replace the functionality.

* Force control-system to use full include paths.

* Remove leftover BatteryMonitor testing files.

* Removes the Simulink Submodule!!!

FrontController is not in a working state yet. Few
more commits until then.

* Fix compiler warnings

Remove defaults AmkManager output. They are managed in the first Update.
Disables ABI change compiler warning/note with -Wno-psabi.
Fix DriverInterface incorrect LUT variable.

* Move Input/Output structs inside classes.

* Template MotorInterface with message concepts.

Removes the AmkBlock's dependency on the generate CAN messages. The
MotorInterface class is defined in terms of the concepts and the FC
`main.cc` can fill in the generated messages later.

* Rename SetPoints concept to disambiguate

* Rename SimpVd and AmkBlock.

Break away from the old Simulink names to be more
clear.
SimpVdInterface --> VehicleDynamics
    renames files and class
AmkBlock --> MotorInterface
    renames file (class already named properly)

* Move shared/controls into FrontController

These functions were only used by VehicleDynamics. Some of the functions
had static variables to hold state. This is bad practice since the
variables are shared over all calls whereas our ControlSystem is Object
Oriented. Converted these functions to classes to hold state in instance
variables.

* add dummy

* board ioc and bindings

* Added majority of update logic for control blocks

* Added more comments for clarity

* Minimal Working FrontController + CLI platform

- Re-implements bindings for the ev5 platform to align with the EV6
names.
- Sends a CAN FC status message (currently always 0) to tell LV that the
system is on.
- Removes FreeRTOS because we don't want to worry about that for now.
We're only running a single task anyways. Reconsider in the future as FC
becomes more complicated.
- This means we could implement the CLI platform which helped verify
that our state machine was working.

* Rename Battery Monitor precharge states.

The `INIT_PRECHARGE` and `PRECHARGE` states had misleading names in the
FSM. The names were 1 state off:
`PRECHARGE` _was_ when all three contactors are closed, but this is the
end of the precharge sequence and the precharge contactor is about to
open to enter the `RUNNING` state.
`INIT_PRECHARGE` _was_ when all NEG & PRE are closed, POS is open. This
is the precharging state.

The labels have been adjusted
`PRECHARGE` -> `PRECHARGE_DONE` and
`INIT_PRECHARGE` -> `PRECHARGE`
to make the code reflect the actual behaviour.

* Prevent immediate transition BM RUNNING state.

When testing the sequence on the physical battery, the precharge
sequence would finish but the contactors would immediately open.

We traced the cause to the "RUNNING" status exit condition which did not
include enough time for the BMS to respond to the contactor command, so
the contactors did not immediately match the desired state and the
system "faulted". Adds a 500ms delay to prevent this.

* Make contactor commands explicit

* TEMPORARY: Bypass pedal & pedal errors for testing

* Fix ContactorFeedback parsing and send FC status

Inverters contactor feedback since it returns 0 for closed and 1 for
open (opposite of the ContactorCommand).

Sends the FCStatus CAN message earlier in the Update since missing AMK
messages used to block the command from running.

* Invert Start Button

* Remove HVIL from FrontController & BatteryMonitor

The HVIL input doesn't exist anymore since the EV5 HSD has been updated.

This commit leaves several tests commented out since we will need them
in the future once we figure out our new HVIL input source.

* Update timing since we slowed it in the bay

* Organize IO and use CubeMX GPIO defines

* Rotate page in datasheet

* Capitalize `Update()` for VehicleDynamics

* Sort signals for easier debugging

---------

Co-authored-by: Samuel Shi <samuelshiwork@gmail.com>
Co-authored-by: luaibash <luai.hawa@gmail.com>