Convert VD, MI to GoogleTest
Removes testing.h since we are completely on GoogleTest now.

Removes test_vehicle_dynamics.cc. The outputs are highly dependent on the
configuration which means out tests will fail if we retune. The low level calc
functions are tested in test_vehicle_dynamics_calc which is a more appropriate place
to test without being susceptible to tuning.