Create stm32f4 project and compile with f7 mcal
Issue #383 in macformula/racecar.
Created the stm32f4 platform for the Demo/Blink project and used the
existing stm32f767 platform mcal to compile it. STM defines macros which
specify which platform is being built. Identifying these macros in the
preprocessor allows for the correct HAL to be included without needing to
edit the mcal when compiling a different platform.