Fix boolean parsing error and simplify generated CAN.
Error was caused by a right bitshift on a `static_cast<bool>` which
was knocking off the value since the cast already sends the bit to the
right. Fixes this by parsing boolean differently than integers.

Removes the `unpack_left/right_shift()` functions and puts the behaviour
directly inside the generated code.

Only applies the scale and offset if the value uses at least one of them.
This simplifies most signals since they have Scale=1 and Offset=0 so no
action is necessary.