Fix cangen boolean unpacking & add tests (#449)
* Set up Cangen C++ Tests. Replicates error in #440

* Fix boolean parsing error and simplify generated CAN.

Error was caused by a right bitshift on a `static_cast<bool>` which
was knocking off the value since the cast already sends the bit to the
right. Fixes this by parsing boolean differently than integers.

Removes the `unpack_left/right_shift()` functions and puts the behaviour
directly inside the generated code.

Only applies the scale and offset if the value uses at least one of them.
This simplifies most signals since they have Scale=1 and Offset=0 so no
action is necessary.

* Fix incorrect test case

* Test Cangen in Github Action

* Test and build firmware on `cangen` change.

Our projects are so dependent on the `cangen` that they should be
tested when it changes.

In particular, changes to `cangen` will trigger the `Demo/TestCangen`
project to verify it generates correct source code.

* Bump cangen to 0.2.0