Implement DAC peripheral for stm32f7
* implements DAC for analog output voltage

* adds HAL_DAC_Start to SetVoltage function

* fixes main.cc demo tests

* Port DAC project to stm32f767

Original project was made for stmf4 which we do not use.
Updates mcal to use the f7 header. Creates the ioc for the nucleo board.

Simplifies the Demo file my converting defines to constants and using
global parameters instead of local arguments to control the min/max
voltage.

The behaviour was verified on an oscilloscope.

* force an error on AnalogDAC stm32f4 since there is no mcal

* restore gpio

* Clarify readme

---------

Co-authored-by: Tamara Xu <xut75@mcmaster.ca>
Co-authored-by: BlakeFreer <freerb@mcmaster.ca>