FrontController tuning parameters.
Separate tuning.hpp and physical.hpp to emphasize that physical
parameters should not be used to tune the handling.

Extracts the pedal_torque curve out of main.cc into tuning.hpp

Simplifies the pedal and steering wheel functions. Replaces the
LinearMap with a custom function since we were spending a lot of effort
just trying to fit into the LinearMap constructor.