Purge the "Mappers" classes (#462)
* Remove unused constant and quadratic map

* Simplify mappers in TMS

Move the lookup tables for TempSensor and FanController into their
classes. No need to have them so accessible.

Replaces LinearMap with direct calculations. It was too much abstraction
to use them.

* remove identity map from LV Fans

* Move const outside of typedef

* FrontController tuning parameters.

Separate tuning.hpp and physical.hpp to emphasize that physical
parameters should not be used to tune the handling.

Extracts the pedal_torque curve out of main.cc into tuning.hpp

Simplifies the pedal and steering wheel functions. Replaces the
LinearMap with a custom function since we were spending a lot of effort
just trying to fit into the LinearMap constructor.

* R.I.P. Mappers (Jan 22, 2024 - Apr 14, 2025)