Added stm32f4 platform on Demo/CAN/Receive project
Issue #383. Stm32f4 platform uses the exact same gpio pin as the f7, but
uses CAN1 instead of CAN3 since the latter is not available on the f4.
Added compile time definitions in respective mcal_conf.cmake files
for each stm32 platform. This is so clangd knows which platform to parse.

In the file mcal/stm32f/periph/can.cc when compiling for f4, it will
throw an error as cubemx does not define CAN3 for this platform. Added a
macro that will define CAN3 as NULL when compiling for stm32f4
platforms. CAN3 should not be reached in the if statement anyways.