Added stm32f4 platform on Demo/CAN/Send project
Issue #383. Stm32f4 platform uses the exact same gpio pin as the f7, but
uses CAN1 instead of CAN3 since the latter is not available on the f4.
Added compile time definitions in respective mcal_conf.cmake files
for each stm32 platform. This is so clangd knows which platform to
parse.
The stm32f4 platform for Demo/CAN/Send has been setup the exact same way
as the f4 for Demo/CAN/Receive.