Guard STM mcal peripheral files
Issue #345. Updated the mcal peripheral files of analog input/output,
gpio, and pwm to only be enabled when they are used by the
platform/project. This prevents clangd from showing a lot of errors
when the dependencies do not exist.
Chose to check for the timer module enabled flag for periph PWM.
The other periphs were straightforward.