Support both F7 and F4 in the same `mcal-stm32f`
Both f-series boards have near identical HALs. This PR lets us define the HAL once, for both platforms, improving maintainability.

* Create stm32f4 project and compile with f7 mcal

Issue #383 in macformula/racecar.
Created the stm32f4 platform for the Demo/Blink project and used the
existing stm32f767 platform mcal to compile it. STM defines macros which
specify which platform is being built. Identifying these macros in the
preprocessor allows for the correct HAL to be included without needing to
edit the mcal when compiling a different platform.

* Added stm32f4 platform for Demo/BasicIO project

Issue #383. Stm32f4 platform uses the exact same pins, configuration,
and bindings as the stm32f767 platform.

* Added stm32f4 platform on Demo/AnalogInput project

Issue #383. Stm32f4 platform uses the exact same pins, configuration,
and bindings as the stm32f767 platform.
Added compile time definitions in respective mcal_conf.cmake files
for each stm32 platform. This is to clangd know which platform to parse.
Included adc rank definition in stm32 mcal when compiling for stm32f4,
as it will not compile without this flag. This solution seemed better
than hardcoding .Rank of 1.

* Format updated stm32f mcal files

The stm32f mcal had been updated to support both stm32f767 and
stm32f4. As it was not formatted to repo standards when the updates were
originally committed, the formatting changes are being pushed in this
commit.

* Add compile definitions to stm32 platforms

Issue #383. Added compile time definitions for the stm32f4 and stm32f767
platforms for the Demo/Blink and Demo/BasicIO projects. This helps
clangd to know which files to parse.

* Changed stm32f767 mcal to stm32f

Issue #383. The stm32f767 mcal folder, cmake library, and namespace
under the /firmware directory have been renamed to stm32f. Since this mcal
can also be used to compile the stm32f4 platform, it does not make sense to
name it stm32f767 specifically.

* Added stm32f4 platform on Demo/CAN/Receive project

Issue #383. Stm32f4 platform uses the exact same gpio pin as the f7, but
uses CAN1 instead of CAN3 since the latter is not available on the f4.
Added compile time definitions in respective mcal_conf.cmake files
for each stm32 platform. This is so clangd knows which platform to parse.

In the file mcal/stm32f/periph/can.cc when compiling for f4, it will
throw an error as cubemx does not define CAN3 for this platform. Added a
macro that will define CAN3 as NULL when compiling for stm32f4
platforms. CAN3 should not be reached in the if statement anyways.

* Added stm32f4 platform on Demo/CAN/Send project

Issue #383. Stm32f4 platform uses the exact same gpio pin as the f7, but
uses CAN1 instead of CAN3 since the latter is not available on the f4.
Added compile time definitions in respective mcal_conf.cmake files
for each stm32 platform. This is so clangd knows which platform to
parse.
The stm32f4 platform for Demo/CAN/Send has been setup the exact same way
as the f4 for Demo/CAN/Receive.

* Guard STM mcal peripheral files

Issue #345. Updated the mcal peripheral files of analog input/output,
gpio, and pwm to only be enabled when they are used by the
platform/project. This prevents clangd from showing a lot of errors
when the dependencies do not exist.
Chose to check for the timer module enabled flag for periph PWM.
The other periphs were straightforward.

* Remove ADC import. The general hal.h file will include it if necessary. This avoids warnings when the project doesn't generate hal_adc.h

* Update remaining projects from stm32f767 to stm32f

* Update docs for generic stm32f configuration

* Fix DAC platform

* Enable CAN interrupts on Demo/CAN stm32f4 and change CAN3 disable mechanism for F4

---------

Co-authored-by: BlakeFreer <freerb@mcmaster.ca>