Hack to avoid dropping CAN Tx messages
STM32 only has 3 hardware Tx mailboxes, so sending 4+ messages in quick
succession is impossible.

Previously we would immediately drop the 4th. This change will attempt
to load it for up to 1 ms.

This is NOT good design since it blocks the task until the queue is
available. Good code would use a FreeRTOS queue to hold all messages and
have a separate high-priority task load the Tx mailbox whenever it is
available. This means that the sending tasks can quickly load its
message and move on.