FrontController rewrite (#481)
* destructure control-system to move to modules

* Rename battery_monitor to accumulator

* Modularize accumulator

* Modularize Governor

* Vehicle Dynamics to Module

- Separate modules for sensors to centralize reading
- Remove the TUNING profile since it won't make 2025 comp
- Fix a CMake include dir bug

* Let each module define its own Command/State enum.

Namespaces keep them separate but consistent.
Motors and driver_interface still need to be converted to modules.

* Add voltage checks to accumulator

* Largely remove DriverInterface

The stateflow just mediated communication between the governor and
dashboard. Most of the states had no effect on the output so
`driver_interface` is now stateless.

* Debugging

* Add accumulator debug message.
Remove optional from main fsm transitions

* Finally removes the Governor!

The governor logic belonged in the main FC state machine. This PR combines the modules to make one continuous state machine.

LV Fans run at 100Hz instead of 1khz
DBC hash is check asynchronously. It does not cause an error but is logged to CAN. Should become an alert later.

Driver Torque Request is passed to `vehicle_dynamics` by the FSM. Lets us externally force it to 0 when not ready to drive. Same with accumulator command.

* Improve Accumulator safety by detecting faults

If the contactor feedback is unexpected, go to a fault state.

This prevents contactors from immediately closing without the boot
sequence after an interrupt (ex TSMS)