Better data handling
Removed type template from MovingAverage. We do not use doubles so this
feature just made it hard to maintain.

Adds a 10-step moving average to the driver pedals and steering wheel.
This is updated at 100Hz so the response delay is 1/2*(10 / 100Hz) = 50
ms.

Adds thresholds to limit motor output. They are enforced when the Amk
request is updated externally.